entity-manager: Add meson option to cache config
For the development use-case configuration is often changed and the
general expectation is for a service to re-parse it's configuration when
it is restarted.

Add a meson option which controls if the configuration (usually in
/var/configuration/system.json) is cached there, or not cached at all.

The new meson option is 'entity-manager-cache-configuration' and by
default it is true, so there is no behavior change when it is not
configured.

Tested: on Tyan S5549 board.
The configuration is cached when the option is left as default and the
configuration is not written to cache when it is set to false.

```
Jan 21 15:50:22 s5549-bmc-a0423f63b4a5 systemd[1]: Starting Entity Manager...
Jan 21 15:50:22 s5549-bmc-a0423f63b4a5 systemd[1]: Started Entity Manager.
Jan 21 15:50:35 s5549-bmc-a0423f63b4a5 entity-manager[32126]: Inventory Added: Supermicro PWS 920P SQ 1
Jan 21 15:50:35 s5549-bmc-a0423f63b4a5 entity-manager[32126]: Inventory Added: MBX 1.60 Chassis
Jan 21 15:50:35 s5549-bmc-a0423f63b4a5 entity-manager[32126]: Inventory Added: Tyan S5549 Baseboard
```

The cache directory now only contains the 'version' file which is
separate from the caching concern, and not affected by this patch.

```
ls -l /var/configuration/
-rw-r--r--    1 root     root            10 Jan 16 11:50 version
```

+++ src/entity_manager.cpp
@@ -144,6 +144,10 @@ static std::shared_ptr<sdbusplus::asio::dbus_interface> createInterface(
 // writes output files to persist data
 bool writeJsonFiles(const nlohmann::json& systemConfiguration)
 {
+    if (!EM_CACHE_CONFIGURATION)
+    {
+        return true;
+    }
     std::filesystem::create_directory(configurationOutDir);
     std::ofstream output(currentConfiguration);
     if (!output.good())
@@ -1311,41 +1315,46 @@ int main()
     });
     tryIfaceInitialize(entityIface);
 
-    if (fwVersionIsSame())
+    if (EM_CACHE_CONFIGURATION)
     {
-        if (std::filesystem::is_regular_file(currentConfiguration))
+        if (fwVersionIsSame())
         {
-            // this file could just be deleted, but it's nice for debug
-            std::filesystem::create_directory(tempConfigDir);
-            std::filesystem::remove(lastConfiguration);
-            std::filesystem::copy(currentConfiguration, lastConfiguration);
-            std::filesystem::remove(currentConfiguration);
-
-            std::ifstream jsonStream(lastConfiguration);
-            if (jsonStream.good())
+            if (std::filesystem::is_regular_file(currentConfiguration))
             {
-                auto data = nlohmann::json::parse(jsonStream, nullptr, false);
-                if (data.is_discarded())
+                // this file could just be deleted, but it's nice for debug
+                std::filesystem::create_directory(tempConfigDir);
+                std::filesystem::remove(lastConfiguration);
+                std::filesystem::copy(currentConfiguration, lastConfiguration);
+                std::filesystem::remove(currentConfiguration);
+
+                std::ifstream jsonStream(lastConfiguration);
+                if (jsonStream.good())
                 {
-                    std::cerr
-                        << "syntax error in " << lastConfiguration << "\n";
+                    auto data =
+                        nlohmann::json::parse(jsonStream, nullptr, false);
+                    if (data.is_discarded())
+                    {
+                        std::cerr
+                            << "syntax error in " << lastConfiguration << "\n";
+                    }
+                    else
+                    {
+                        lastJson = std::move(data);
+                    }
                 }
                 else
                 {
-                    lastJson = std::move(data);
+                    std::cerr << "unable to open " << lastConfiguration << "\n";
                 }
             }
-            else
-            {
-                std::cerr << "unable to open " << lastConfiguration << "\n";
-            }
         }
-    }
-    else
-    {
-        // not an error, just logging at this level to make it in the journal
-        std::cerr << "Clearing previous configuration\n";
-        std::filesystem::remove(currentConfiguration);
+        else
+        {
+            // not an error, just logging at this level to make it in the
+            // journal
+            std::cerr << "Clearing previous configuration\n";
+            std::filesystem::remove(currentConfiguration);
+        }
     }
 
     // some boards only show up after power is on, we want to not say they are
